Expand support for OpenAPI 3.1.
Support items schema being absent under arrays
