+++ src/test/resources/examples/openapi310/models/ContainsUntypedArray.kt
@@ -1,0 +1,13 @@
+package examples.openapi310.models
+
+import com.fasterxml.jackson.`annotation`.JsonProperty
+import javax.validation.constraints.NotNull
+import kotlin.Any
+import kotlin.collections.List
+
+public data class ContainsUntypedArray(
+  @param:JsonProperty("untyped_array")
+  @get:JsonProperty("untyped_array")
+  @get:NotNull
+  public val untypedArray: List<Any?>,
+)
